\name{bypass}
\docType{data}
\alias{bypass}
\title{Gene Expression Profiling of the Human Adipose Tissue of Massively Obese Subjects Before 
            and After Bariatric Surgery}
\description{
            This data set resulted from a study investigating the dynamics of adipose tissue 
            gene expression profiles induced in massively obese subjects by a form of 
            bariatric surgery (gastric bypass). Please see the references provided below for 
            further experimental details.
          }
\usage{data(bypass)}
\source{\url{http://corneliu.henegar.info/FunNet.htm}}
\format{Three data frames containing significant up-regulated or down-regulated genes and a reference list 
	    with all the genes tested for differential expression during this experiment.}
\references{
            1. Cancello R, Henegar C, Viguerie N, Taleb S, Poitou C, Rouault C, Coupaye M, Pelloux V, Hugol D, 
            Bouillot JL, Bouloumie A, Barbatelli G, Cinti S, Svensson PA, Barsh GS, Zucker JD, Basdevant A, 
            Langin D, Clement K. Reduction of macrophage infiltration and chemoattractant gene expression 
            changes in white adipose tissue of morbidly obese subjects after surgery-induced weight loss. 
            Diabetes 2005; 54(8):2277-86.
            
            2. Henegar C, Tordjman J, Achard V, Lacasa D, Cremer I, Guerre-Millo M, Poitou C, Basdevant A,		
	    Stich V, Viguerie N, Langin D, Bedossa P, Zucker J-D, Clement K. Adipose tissue transcriptomic 
	    signature highlights the pathologic relevance of extracellular matrix in human obesity. In positive
	    revision in Genome Biology (October 2007).
	                
            3. GEO accession number: GSE9007. \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE9007}.
           }
\seealso{
            \code{\link{FunNet}}.
}
\keyword{datasets}