\name{obese}
\docType{data}
\alias{obese}
\title{Gene Expression Profiling of the Human White Adipose Tissue in Obese Subjects}
\description{
            This data set resulted from microarray experiments performed on human white adipose 
            tissue comparing the expression profiles in obese subjects and lean controls. Please 
            see the references provided below for further experimental details.
          }
\usage{data(obese)}
\source{\url{http://corneliu.henegar.info/FunNet.htm}}
\format{Three data frames containing significant genes up-regulated or down-regulated in human WAT of obese subjects
            and a reference list with all the genes tested for differential expression during this experiment.}
\references{
            1. Clement K, Viguerie N, Poitou C, Carette C, Pelloux V, Curat CA, Sicard A, Rome S, 
            Benis  A,  Zucker  JD  et  al:  Weight  loss  regulates  inflammation-related  genes  in 
            white adipose tissue of obese subjects. FASEB J 2004, 18(14):1657-1669.
            
            2. Henegar C, Tordjman J, Achard V, Lacasa D, Cremer I, Guerre-Millo M, Poitou C, Basdevant A,		
	    Stich V, Viguerie N, Langin D, Bedossa P, Zucker J-D, Clement K. Adipose tissue transcriptomic 
	    signature highlights the pathologic relevance of extracellular matrix in human obesity. In positive
	    revision in Genome Biology (October 2007).
	                
            3. GEO accession number: GSE9153. \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE9153}.
           }
\seealso{
            \code{\link{FunNet}}.
}
\keyword{datasets}