\name{RN.GO.DIR.CC.file.annot}
\alias{RN.GO.DIR.CC.file.annot}
\docType{data}
\title{Internal routine or variable}
\description{
  This internal routine or variable should not be accessed or modified directly by the user.
}
\seealso{
           \code{\link{FunNet}, \link{annotations}}.
}
\keyword{internal}
