\name{yeast.cycle}
\docType{data}
\alias{yeast.cycle}
\title{Yeast Cell Cycle Expression Profiling}
\description{
            This data set resulted from microarray experiments performed on yeast 
	    (Saccharomyces cerevisiae) to study the dynamics of cell cycle related processes in yeast. 
	    Please see the reference provided below for further experimental details.
          }
\usage{data(yeast.cycle)}
\source{\url{http://corneliu.henegar.info/FunNet.htm}}
\format{A data frame containing temporal series gene expression data obtained from yeast (Saccharomyces 
            cerevisiae) cultures performed to explore the cell cycle.}
\references{
            Spellman P, Sherlock G, Zhang M, Iyer V, Anders K, Eisen M, Brown P, Botstein D, 
	    Futcher B. Comprehensive identification of cell cycle-regulated genes of the yeast Saccharomyces 
            cerevisiae by microarray hybridization. Mol Biol Cell 1998; 9(12):3273-97.
           }
\seealso{
            \code{\link{FunNet}}.
}
\keyword{datasets}