\name{yeast.dna}
\docType{data}
\alias{yeast.dna}
\title{Expression Profiling of the Yeast after Irradiation}
\description{
            This data set resulted from microarray experiments performed on irradiated yeast 
            (Saccharomyces cerevisiae) to study the dynamics of DNA repairing processes induced by the 
            damages inflicted by ionizing radiation in yeast. Please see the reference provided 
            below for further experimental details.
          }
\usage{data(yeast.dna)}
\source{\url{http://corneliu.henegar.info/FunNet.htm}}
\format{A data frame containing temporal series gene expression data obtained from irradiated yeast (Saccharomyces 
            cerevisiae).}
\references{
            Gasch A, Huang M, Metzner S, Botstein D, Elledge S, Brown P. Genomic expression responses to DNA-damaging
            agents and the regulatory role of the yeast ATR homolog Mec1p. Mol Biol Cell 2001; 12(10):2987-3003.
           }
\seealso{
            \code{\link{FunNet}}.
}
\keyword{datasets}
